Major changes to bootstrap (again).
The default location of bash in the hashbang line was wrong, but it's fixed now.
Cut out all the quieting flags, since it isn't quite robust enough to be trusted.
Switched from wget to cURL so the downloads don't flood STDIN on vagrant setup.
Removed all the cd's I could, so it basically sits still.
Added a modification to a config file that apparently need to have an absolute path.